separating populate scripts
