+++ scripts/populate/utils/toBytes32.mjs
@@ -1,0 +1,3 @@
+import { keccak256, toUtf8Bytes } from 'ethers'
+
+export const toBytes32 = (str) => keccak256(toUtf8Bytes(str))
