+++ scripts/populate/utils/whenCalledDirectly.mjs
@@ -1,0 +1,5 @@
+export const whenCalledDirectly = async (handler) => {
+  if (process.argv.includes('scripts/populate/index.mjs') === false) {
+    await handler()
+  }
+}
